Split native messaging into renderer and background service. Encrypt messages and verify timestamp
